requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap==0.2.4
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.3.0
